Add tracewriter logging for statistics functions
