+++ src/SFA.DAS.Data.AcceptanceTests/Functions/Statistics/Commitments/WhenIGetCommitmentsStatistics.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using SFA.DAS.Data.Domain.Models.Statistics.Commitments;
@@ -19,7 +20,9 @@ namespace SFA.DAS.Data.AcceptanceTests.Functions.Statistics.Commitments
             };
             CommitmentsStatisticsHandlerMock.Setup(x => x.Handle()).Returns(Task.FromResult(commitmentsExternalModel));
 
-            await DAS.Data.Functions.Statistics.GetCommitmentStatisticsFunction.Run(null, Log, StatisticsService);
+            var logger = new TraceWriterStub(TraceLevel.Verbose);
+
+            await DAS.Data.Functions.Statistics.GetCommitmentStatisticsFunction.Run(null, logger, Log, StatisticsService);
 
             var databaseAsExpected = TestHelper.ConditionMet(IsDatabaseInExpectedState, TimeSpan.FromSeconds(60));
