+++ src/SFA.DAS.Data.AcceptanceTests/SFA.DAS.Data.AcceptanceTests.csproj
@@ -334,6 +334,7 @@
     <Compile Include="HmrcDataLoadTests\WhenDataLoadIsExecutedWithNoPendingImports.cs" />
     <Compile Include="HmrcDataLoadTests\WhenDataLoadIsExecutedWithNoValidationTests.cs" />
     <Compile Include="HmrcDataLoadTests\WhenDataLoadIsExecutedWithNullCessationDates.cs" />
+    <Compile Include="TraceWriterStub.cs" />
     <Compile Include="TransferEventTests\WhenAPeriodEndEventIsCreated.cs" />
     <Compile Include="PaymentEventTests\PaymentEventTestsBase.cs" />
     <Compile Include="PaymentEventTests\WhenAPeriodEndEventIsCreated.cs" />
