+++ src/SFA.DAS.Data.Functions.UnitTests/packages.config
@@ -4,8 +4,8 @@
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net462" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net462" />
   <package id="Microsoft.Azure.KeyVault.Core" version="2.0.4" targetFramework="net462" />
-  <package id="Microsoft.Azure.WebJobs" version="2.1.0" targetFramework="net462" />
-  <package id="Microsoft.Azure.WebJobs.Core" version="2.1.0" targetFramework="net462" />
+  <package id="Microsoft.Azure.WebJobs" version="2.2.0" targetFramework="net462" />
+  <package id="Microsoft.Azure.WebJobs.Core" version="2.2.0" targetFramework="net462" />
   <package id="Microsoft.Azure.WebJobs.Extensions" version="2.1.0" targetFramework="net462" />
   <package id="Microsoft.Data.Edm" version="5.8.2" targetFramework="net462" />
   <package id="Microsoft.Data.OData" version="5.8.2" targetFramework="net462" />
