+++ src/SFA.DAS.Data.Functions/Statistics/GetAccountStatisticsFunction.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Host;
 using SFA.DAS.Data.Domain.Interfaces;
 using SFA.DAS.Data.Functions.Ioc;
 using SFA.DAS.NLog.Logger;
@@ -9,13 +11,19 @@ namespace SFA.DAS.Data.Functions.Statistics
     public static class GetAccountStatisticsFunction
     {
         [FunctionName("GetAccountStatisticsFunction")]
-        public static async Task Run([TimerTrigger("%CronSchedule%")] TimerInfo myTimer, [Inject] ILog log,
+        public static async Task Run([TimerTrigger("%CronSchedule%")] TimerInfo myTimer, TraceWriter traceLog, [Inject] ILog log,
             [Inject] IStatisticsService statsService)
         {
             log.Debug("Gathering statistics for the EAS area of the system");
-            var message = await statsService.CollateEasMetrics();
 
-           
+            try
+            {
+                var message = await statsService.CollateEasMetrics();
+            }
+            catch (Exception e)
+            {
+                traceLog.Error($"Error processing GetAccountStatisticsFunction to pull data into RDS: {e.Message}", e);
+            }
         }
     }
 }
