+++ src/SFA.DAS.Data.Functions/Statistics/GetPaymentsStatisticsFunction.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Host;
 using SFA.DAS.Data.Domain.Interfaces;
 using SFA.DAS.Data.Functions.Ioc;
 using SFA.DAS.NLog.Logger;
@@ -10,12 +12,20 @@ namespace SFA.DAS.Data.Functions.Statistics
     {
         [FunctionName("GetPaymentsStatisticsFunction")]
        // [Disable]
-        public static async Task Run([TimerTrigger("%CronSchedule%")] TimerInfo myTimer, [Inject] ILog log,
+        public static async Task Run([TimerTrigger("%CronSchedule%")] TimerInfo myTimer, TraceWriter traceLog, [Inject] ILog log,
         [Inject] IStatisticsService statsService)
         {
             log.Info("Gathering statics for the payments area of the system");
 
-            await statsService.CollatePaymentStatisticsMetrics();
+            try
+            {
+                await statsService.CollatePaymentStatisticsMetrics();
+            }
+            catch (Exception e)
+            {
+                traceLog.Error($"Error processing GetPaymentsStatisticsFunction to pull data into RDS: {e.Message}", e);
+                throw;
+            }
         }
     }
 }
